requirements.txt
python-telegram-bot==21.9
python-dotenv==1.0.0
httpx[socks]>=0.27.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pytest>=8.0.0
